REFACTOR: 클래스 통합 및 정리 #1
