+++ src/main/java/com/dwu/alonealong/controller/product/CartController.java
@@ -1,7 +1,8 @@
-package com.dwu.alonealong.controller;
+package com.dwu.alonealong.controller.product;
 
 import javax.servlet.http.HttpServletRequest;
 
+import com.dwu.alonealong.controller.UserSession;
 import com.dwu.alonealong.exception.NullProductException;
 import com.dwu.alonealong.exception.StockException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,14 +18,49 @@ import com.dwu.alonealong.domain.Product;
 import com.dwu.alonealong.service.AloneAlongFacade;
 import com.dwu.alonealong.exception.UserNotMatchException;
 
+import java.util.List;
+
 @Controller
 @SessionAttributes({"userSession"})
-public class CartItemController {
+public class CartController {
 	private AloneAlongFacade aloneAlong;
 
 	@Autowired
 	public void setAloneAlong(AloneAlongFacade aloneAlong) {
 		this.aloneAlong = aloneAlong;
+	}
+
+	@RequestMapping("/cart")
+	public String viewCart(HttpServletRequest request,
+								@RequestParam(value="productId", required=false) String productId,
+								@RequestParam(value="stockError", required=false) boolean stockError,
+								ModelMap model) throws Exception {
+		UserSession userSession = (UserSession)request.getSession().getAttribute("userSession");
+		if(userSession == null) {
+			return "forward:/login";
+		}
+		String userId = userSession.getUser().getId();
+
+		int totalPrice = 0;
+		int shippingFee = 0;
+		List<CartItem> cart = this.aloneAlong.getAllCartItem(userId);
+		for(CartItem cartItem : cart) {
+			totalPrice += cartItem.getUnitPrice();
+		}
+
+		model.put("productsPrice", totalPrice);
+		if (totalPrice != 0 && totalPrice <  Product.FREE_SHIPPING_PRICE) {
+			totalPrice += Product.SHIPPING_FEE;
+		}
+		model.put("shippingFee", shippingFee);
+		model.put("totalPrice", totalPrice);
+		model.put("cart", cart);
+
+		if(stockError == true) {
+			model.put("cartItemName", aloneAlong.getProduct(productId).getProductName());
+		}
+		return "thyme/Cart";
+//		return "productCart";
 	}
 
 	@RequestMapping("/cart/delete")
